V2.3.1 - rev.28 - new permission "Event Editor"; more readable browser info; fixed user deletion of statistics - statistics: interpretation of browser info (more readable) - statistics: only list browser info, if it is not empty (!='') - statistics: fixed deletion of user statistics - settings: new user permission "Event Editor",   possibility to only edit details of an existing event - userView: added "Event Editor" to listing of permissions per user - userView: also highlight users with Event Editor permissions - index: show icon "edit event" for Event Editors - edit event: Event Editor now can edit event relevant data - edit event: text representation of non-editable data               is formatted bold (strong) now
internal stuff:
- new function: isEventEditor

+++ trunk/statistics.php
@@ -17,12 +17,64 @@ if (!isAdmin()) {
 	exit;
 }
 
+function detect($user_agent) 
+{	
+	////--------------------------------------------------------------------
+	////code by blpgirl (Leyla Maria Bonilla Palacio, Columbia, http://loquelediga.com/) - found on http://snipplr.com/view/35381/
+	////which references class 'browserdetect' by Paul Scott, South Africa - on http://www.phpclasses.org/browse/package/2827.html
+	////allowed to use under Terms of GPL
+	////--------------------------------------------------------------------
+	
+    //nota: firefox debe ir luego de mozilla pues el user agent tiene ambos
+    //y flock debe ir luego de mozilla y firefox por la misma razon
+    //chrome también debe ir despues de mozilla y safari porque los contiene a ambos
+    //funciona porque siempre queda el valor de la última cadena encontrada
+    $browser = array ("IE","OPERA","MOZILLA","NETSCAPE","FIREFOX","SAFARI", "FLOCK", "CHROME");
+    $os = array ("WINDOWS","MAC","IPHONE","IPAD","ANDROID"); //// modiefied, was: ("WINDOWS","MAC","IPHONE");
+    $info['browser'] = "OTHER";
+    $info['os'] = "OTHER";
+    ////$user_agent = $_SERVER['HTTP_USER_AGENT'];  ////modiefied, is given now by parameter
+    //Por cada valor del array de navegadores
+    foreach ($browser as $parent)
+    {
+		//con strtoupper devuelve la cadena en mayúsculas y con strpos devuelve la posicion de la cadena
+		//Si no se encuentra la cadena, devuelve FALSE
+		$s = strpos(strtoupper($user_agent), $parent);
+		//el user agent siempre suelta: Navegador/NumVersion o Navegador NumVersion pa explorer
+		//con esto se tiene la posicion para la version que es justo despues del navegador (s + tamaño nombre navegador)
+		$f = $s + strlen($parent);
+		//devuelve la cadena que empieza en el caracter f y termina en f+5
+		$version = substr($user_agent, $f, 5);
+		//reemplaza el numero, punto o / en la cadena version y lo reemplaza por vacio
+		$version = preg_replace('/[^0-9,.]/','',$version);
+		if (!($s===false)) ////modified, was: if ($s)
+		{
+			//como se encontro el navegador, se asignan los valores
+			$info['browser'] = $parent;
+			$info['version'] = $version;
+		}
+    }
+    foreach ($os as $val)
+    {
+		//eregi encuentra subcadenas sin diferenciar mayusculas de minusculas
+		if (eregi($val,strtoupper($user_agent))) $info['os'] = $val;
+    }
+	return $info;
+}
+
+
 // Get the query string
 $method = $_GET["method"];
 
 // If the form has been submitted, then we need to handle the data.
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 	if($method == 'truncate') {
+		$sql = "CREATE TABLE tmp_system_statistics as SELECT * from cr_statistics WHERE type='system'";
+		if (!mysql_query($sql))
+ 	 		{
+  			die('Error: ' . mysql_error());
+  			}
+		
 		$sql = ("TRUNCATE TABLE cr_statistics");
 		if (!mysql_query($sql))
  	 		{
@@ -34,7 +86,23 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
  	 		{
   			die('Error: ' . mysql_error());
   			}
+
+		$sql = "INSERT INTO cr_statistics (userid,date,type,detail1,detail2,detail3,script) ";
+		$sql = $sql . "SELECT userid,date,type,detail1,detail2,detail3,script from tmp_system_statistics order by date";
+		if (!mysql_query($sql))
+ 	 		{
+  			die('Error: ' . mysql_error());
+  			}
+
+		$sql = "DROP TABLE tmp_system_statistics";
+		if (!mysql_query($sql))
+ 	 		{
+  			die('Error: ' . mysql_error());
+  			}
+
 		insertStatistics("system",__FILE__,"statistics deleted");
+
+
 		
 		// After we have truncated the data, we want to reload the page
 		header('Location: statistics.php'); // Move to the home page of the admin section
@@ -70,12 +138,19 @@ include('includes/header.php');
 				echo "</thead>";
 				echo "<tbody>";
 				
-					$sql = "SELECT detail3 as browser, count(*) as count from cr_statistics where detail1 like 'login%' group by detail3 order by count desc ".$browserLimit;
+					$sql = "SELECT detail3 as browser, count(*) as count from cr_statistics where detail1 like 'login%' and detail3!='' group by detail3 order by count desc ".$browserLimit;
 					$result = mysql_query($sql) or die(mysql_error());
 					while($row = mysql_fetch_array($result, MYSQL_ASSOC)) { 
 						extract($row);
 						echo "<tr>";
-						echo "<td>".$browser."</td>";
+						//echo "<td>".$browser."</td>";
+
+						//$arrBrowser = get_browser($browser,true)
+						//echo "<td>".$arrBrowser['parent']." on ".$arrBrowser['platform']."</td>";
+						
+						//http://snipplr.com/view/35381/detectar-browser/
+						$d = detect($browser);
+						echo "<td>".$d['browser']." ".$d['version']." on ".$d['os']."</td>";
 						echo "<td>".$count."</td>";
 						echo "</tr>";
 					}
@@ -111,12 +186,13 @@ include('includes/header.php');
 		</tbody>
 		</table>
 		
-		<a href="#" data-reveal-id="truncStatData" class="button">Delete Statistics</a>
+		<a href="#" data-reveal-id="truncStatData" class="button">Delete User Statistics</a>
 
 			<div id="truncStatData" class="reveal-modal">
-     			<h1>Really delete all statistics?</h1>
+     			<h1>Really delete user statistics?</h1>
 				<p>Are you sure you really want to delete ALL user statistics data? <br>There is no way of undoing this action.</p>
-				<p><a class="button" href="statistics.php?methode=truncate">Sure, delete statistics</a></p>
+				<p><form action="statistics.php?method=truncate" method="post" id="truncate">
+				<input type="submit" value="Sure, delete statistics" /></form></p>
      			<a class="close-reveal-modal">&#215;</a>
 			</div>
